+++ orm/BookshelfMapping.ts
@@ -43,6 +43,7 @@ export default class BookshelfMapping {
     public writeableSqlColumns: IWritableSqlColumnDescriptor[];
     public readableSqlColumns: IReadableSqlColumnDescriptor[];
     public qualifiedRegularColumnNames: string[];
+    public isNew?: () => boolean;
 
     public constructor(dbContext: Bookshelf, config: IDbMapping) {
         this.dbContext = dbContext;
@@ -56,6 +57,7 @@ export default class BookshelfMapping {
         this.keepHistory = BookshelfMapping.getOptionOrDefault(config.keepHistory, false);
         this.historyColumns = BookshelfMapping.getOptionOrDefault(config.historyColumns, { revisionId: "revision_id", parentId: "parent_id" });
         this.historyChangeCheck = BookshelfMapping.getOptionOrDefault(config.historyChangeCheck, false);
+        this.isNew = config.isNew;
 
         this.configureHistory();
 
@@ -113,10 +115,14 @@ export default class BookshelfMapping {
     }
 
     private createModel(): typeof Bookshelf.Model {
-        const prototype = {
+        const prototype: { tableName: string; idAttribute: string; isNew?: () => boolean } = {
             tableName: this.tableName,
             idAttribute: this.identifiedBy
         };
+
+        if (this.isNew) {
+            prototype.isNew = this.isNew;
+        }
 
         this.relations.forEach(this.addRelation.bind(this, prototype));
         return this.dbContext.Model.extend(prototype) as any;